Start Refacting Collapser Algorithm
- Use Timer class
- Use Priority Queue instead of randFloor() function
The Refactor will be done in stages, like how the collapsing algorithm will now work. I commented out the code for collapsing a tile not on the path and fatal collapsing, and only worked to fix collapsing a tile not between the player and the exit. I will work on the former for the next commit.