Stop tracking floors in TileMap
There is a bug that causes the FloorPicker to sometimes run out of tiles
before the game has started. I have started to fix that bug. The first
step is reducing the amount of state between TileMap, Collapser, and
FloorPicker. TileMap no longer has a floors attribute that tracks all
the floor tiles that are present. This will also make the TileMap class
and collapsing algorithm clearer.